Use string based tokenizer in prefix cache (#774)
* Use string based tokenizer
* add interface for tokenizer
* add string tokenizer test
* nit fixes
* fix lint error
* update hash_test.go

---------

+++ pkg/plugins/gateway/prefixcacheindexer/indexer.go
@@ -25,10 +25,10 @@ import (
 type PrefixCacheIndexer interface {
 	// MatchPrefix matches the longest prefix sequence for input request (passed as input tokens)
 	// and returns matched prefix (as tokens), remaining unmatched input request (as tokens) and pods matching the prefix
-	MatchPrefix(inputTokens []int, model string, pods []*v1.Pod) (matchedTokens []int, unMatchedTokens []int, matchedPods []*v1.Pod)
+	MatchPrefix(inputTokens []byte, model string, pods []*v1.Pod) (matchedTokens []byte, unMatchedTokens []byte, matchedPods []*v1.Pod)
 
 	// AddPrefix adds tokens in internal prefix cache indexer to be used by future requests
-	AddPrefix(tokens []int, model, pod string)
+	AddPrefix(tokens []byte, model, pod string)
 
 	// Evict is invoked at fixed internal to clean up expired tokens from prefix cache.
 	// TODO: Add max blocks to cache, add LRU policy along with TTL and add performance benchmark tests.
